Login and Security

+++ src/main/java/com/bathanh/apibook/domain/book/BookService.java
@@ -1,5 +1,6 @@
 package com.bathanh.apibook.domain.book;
 
+import com.bathanh.apibook.domain.auths.AuthsProvider;
 import com.bathanh.apibook.persistence.book.BookStore;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,12 +13,15 @@ import java.util.UUID;
 import static com.bathanh.apibook.domain.book.BookError.supplyBookAlreadyExist;
 import static com.bathanh.apibook.domain.book.BookError.supplyBookNotFound;
 import static com.bathanh.apibook.domain.book.BookValidation.validate;
+import static com.bathanh.apibook.error.CommonError.supplyForbiddenError;
 
 @Service
 @RequiredArgsConstructor
 public class BookService {
 
     private final BookStore bookStore;
+
+    private final AuthsProvider authsProvider;
 
     public List<Book> findAll() {
         return bookStore.findAll();
@@ -36,6 +40,7 @@ public class BookService {
         validate(book);
         verifyTitleAndAuthorAvailable(book.getTitle(), book.getAuthor());
 
+        book.setUserId(authsProvider.getCurrentUserId());
         book.setCreatedAt(Instant.now());
         return bookStore.create(book);
     }
@@ -43,6 +48,8 @@ public class BookService {
     public Book update(final UUID id, final Book book) {
         validate(book);
         final Book updatedBook = findById(id);
+
+        verifyUpdateBookPermission(updatedBook);
 
         updatedBook.setTitle(book.getTitle());
         updatedBook.setAuthor(book.getAuthor());
@@ -54,7 +61,9 @@ public class BookService {
     }
 
     public void delete(final UUID id) {
-        findById(id);
+        final Book book = findById(id);
+        verifyDeleteBookPermission(book);
+
         bookStore.delete(id);
     }
 
@@ -62,6 +71,20 @@ public class BookService {
         final Optional<Book> bookOptional = bookStore.findByTitleAndAuthor(title, author);
         if (bookOptional.isPresent()) {
             throw supplyBookAlreadyExist(title, author).get();
+        }
+    }
+
+    private void verifyUpdateBookPermission(final Book book) {
+        if (authsProvider.getCurrentUserRole().equals("ROLE_CONTRIBUTOR")
+                && !authsProvider.getCurrentUserId().equals(book.getUserId())) {
+            throw supplyForbiddenError().get();
+        }
+    }
+
+    private void verifyDeleteBookPermission(final Book book) {
+        if (authsProvider.getCurrentUserRole().equals("ROLE_CONTRIBUTOR")
+                && !authsProvider.getCurrentUserId().equals(book.getUserId())) {
+            throw supplyForbiddenError().get();
         }
     }
 }
